==> code/06_Pathway_Analysis.qmd <==
---
title: "Pathway Analyses"
format: html
---

```{r}

library(dplyr)
library(clusterProfiler)
library(org.Hs.eg.db)

```

# Load Data
```{r}

#normalized expression data
source("functions/Load_Normalized_Expression.R")
normalized_data = load_normalized_expression("../outputs/normalized_expression.csv")
rm(load_normalized_expression)

#differential expression data
diff_exp_table = read.csv("../outputs/differential_expression.csv")
rownames(diff_exp_table) = diff_exp_table$X
diff_exp_table = diff_exp_table[,2:ncol(diff_exp_table)]

#sample metadata
source("functions/Load_Sample_Metadata.R")
sample_metadata = load_sample_metadata("../data/sample_metadata.csv")
rm(load_sample_metadata)

```



# DEG Enrichment Pathway Analysis
```{r}

significant_genes = diff_exp_table |> 
  dplyr::filter(adj.P.Val <= 0.05 & abs(logFC) > 1.5)

enrich_GO_results = enrichGO(significant_genes$entrez_symbol,
                             OrgDb = org.Hs.eg.db,
                             ont = "ALL")

```

## Plots
```{r}

clusterProfiler::dotplot(enrich_GO_results, 
                         showCategory = 20)

```


# GSEA
```{r}

#rank genes
ranked_genes = diff_exp_table |> 
  arrange(desc(logFC))
#make into a named vector
ranked_genes = setNames(ranked_genes$logFC, ranked_genes$entrez_symbol)

#GSEA
GESA_KEGG_results = clusterProfiler::gseKEGG(ranked_genes)
GESA_GO_results = clusterProfiler::gseGO(ranked_genes, OrgDb = org.Hs.eg.db)



```

## Plots
```{r}

clusterProfiler::dotplot(GESA_KEGG_results, showCategory = 10, split = ".sign") + facet_grid(.~.sign)
clusterProfiler::dotplot(GESA_GO_results, showCategory = 10, split = ".sign") + facet_grid(.~.sign)

ridgeplot(GESA_KEGG_results, showCategory = 10)
ridgeplot(GESA_GO_results, showCategory = 10)

```

### Save plots
```{r}

plot1 = clusterProfiler::dotplot(GESA_KEGG_results, 
                                 showCategory = 5, 
                                 split = ".sign") + 
  facet_grid(.~.sign)

plot2 = ridgeplot(GESA_GO_results, showCategory = 10)

png(filename = "../outputs/plots/GSEA.png", height = 800, width = 850, res = 100)
ggarrange(plot1, plot2,
          nrow = 2,
          labels = c("GSEA - KEGG", "GSEA - GO"))
dev.off()

```
